Fixed data type of osync_mapping_engine_member_change()'s member id parameter, whichi is long long int - not int.
refs #974



git-svn-id: http://svn.opensync.org/trunk@5455 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e

+++ opensync/engine/opensync_mapping_engine.c
@@ -560,7 +560,7 @@ OSyncChange *osync_mapping_engine_nth_change(OSyncMappingEngine *engine, unsigne
 	return NULL;
 }
 
-OSyncChange *osync_mapping_engine_member_change(OSyncMappingEngine *engine, int memberid)
+OSyncChange *osync_mapping_engine_member_change(OSyncMappingEngine *engine, long long int memberid)
 {
 	OSyncList *e = NULL;
 	osync_assert(engine);
